test(runtime): split Codex L1 integration test
